trying to fix network tests

+++ sdk/include/memory/memory-signaled.hpp
@@ -55,7 +55,7 @@ namespace adam
 
         /** @brief Wait for the signal activation. Compatible with std::chrono::duration */
         template <typename rep, typename period>
-        bool wait(std::chrono::duration<rep, period> timeout = std::chrono::duration<rep, period>::max()) { return wait(timeout == std::chrono::duration<rep, period>::max() ? -1 : std::chrono::duration_cast<std::chrono::milliseconds>(timeout).count()); }
+        bool wait(std::chrono::duration<rep, period> timeout = (std::chrono::duration<rep, period>::max)()) { return wait(timeout == (std::chrono::duration<rep, period>::max)() ? -1 : static_cast<int32_t>(std::chrono::duration_cast<std::chrono::milliseconds>(timeout).count())); }
 
     protected:
